Connect ChangeListener and force initial update on editor creation
Calling changeListenerCallback() once in PluginEditor constructor ensures that GUI will display the number of times PluginProcessor::audioWorkgroupContextChanged had been called by the host before the editor was even created. In Logic Pro X, this reveals that 1 call had occurred.

Adding the addChangeListener() call [it was a mistake to omit this] would ensure that any workgroup-context changes which occur while the editor is open would also be correctly displayed. This does not appear to happen in Logic Pro X, even when the audio I/O devices are changed in Settings. It may be that Logic actually re-instantiates all plug-ins when this happens.